+++ packages/api-reference/src/standalone/index.ts
@@ -1,0 +1,1 @@
+export { createApiReference } from './lib/html-api'
